add periodical task for delete requests

+++ backend/tasks.py
@@ -5,8 +5,9 @@ from requests.cutom_image_handler import ImageHandler
 import zipfile
 from io import BytesIO
 
-from backend.celery import app
+from django.utils import timezone
 
+from backend.celery import app
 
 @app.task
 def task_image_edit(file_id):
@@ -31,26 +32,30 @@ def task_to_zip(file_ids):
             archive.writestr(file.file.name, buffer_file)
     buffer_archive.seek(0)
     
-    file_1 = EditedFile.get_by_id(file_ids[0])
+    file_example = EditedFile.get_by_id(file_ids[0])
     
-    file_1.request.update_file(str(file_1.request.id) + '.zip', buffer_archive.getvalue())
-    file_1.request.update_status_done()
+    file_example.request.update_file(str(file_example.request.id) + '.zip', buffer_archive.getvalue())
+    file_example.request.update_status_done()
+    
+    file_example.request.update_expiration_date()
 
     return True
 
 @app.task
-def task_clear_request(request_id):
-    request = Request.objects.get(id=request_id)
-    
-    uploaded_files = UploadedFile.objects.filter(request=request)
-    for uploaded_file in uploaded_files:
-        uploaded_file.delete_file()
+def task_clear_requests():
+    print(timezone.now())
+    requests = Request.objects.all()
+    for request in requests:
+        if request.expiration_date < timezone.now():
+            uploaded_files = UploadedFile.objects.filter(request=request)
+            for uploaded_file in uploaded_files:
+                uploaded_file.delete_file()
+
+            edited_files = EditedFile.objects.filter(request=request)
+            for edited_file in edited_files:
+                edited_file.delete_file()
+                
+            request.delete_file()
+            request.delete()
         
-    edited_files = EditedFile.objects.filter(request=request)
-    for edited_file in edited_files:
-        edited_file.delete_file()
-        
-    request.delete_file()
-    request.delete()
-    
     return True
